separate add external id modal functionality out from curator; fix references
